msg serialization WIP
