+++ rmm/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "rmm"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.66"
+bincode = "1.3.3"
+derive_more = "0.99.17"
+serde = { version = "1.0.147", features = ["derive"] }
+serde_json = "1.0.87"
+thiserror = "1.0.37"
+uniffi = "0.21.0"
+uniffi_macros = "0.21.0"
+uuid = { version = "1.2.1", features = ["v4", "wasm-bindgen", "js"] }
