+++ rmm/src/cmd.rs
