+++ rmm/src/platform.rs
