+++ shared/Cargo.toml
@@ -9,7 +9,10 @@ name = "shared"
 
 [dependencies]
 anyhow = "1.0.66"
+bincode = "1.3.3"
 derive_more = "0.99.17"
+rmm = { path = "../rmm" }
+shared_types = { path = "../shared_types" }
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = "1.0.87"
 thiserror = "1.0.37"
@@ -18,4 +21,7 @@ uniffi_macros = "0.21.0"
 uuid = { version = "1.2.1", features = ["v4", "wasm-bindgen", "js"] }
 
 [build-dependencies]
+shared_types = { path = "../shared_types" }
+serde-generate = "0.24.0"
+serde-reflection = "0.3.6"
 uniffi_build = { version = "0.21.0", features = ["builtin-bindgen"] }
